+++ src/main/java/ru/javarush/november/zaharin/entities/Answer.java
@@ -1,4 +1,4 @@
-package ru.javarush.november.zaharin.thegame.entity;
+package ru.javarush.november.zaharin.entities;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
@@ -60,5 +60,14 @@ public class Answer {
     @Override
     public int hashCode() {
         return Objects.hash(answerText, id, nextQuestionId);
+    }
+
+    @Override
+    public String toString() {
+        return "Answer{" +
+                "answerText='" + answerText + '\'' +
+                ", id=" + id +
+                ", nextQuestionId=" + nextQuestionId +
+                '}';
     }
 }
