+++ src/main/java/ru/javarush/november/zaharin/services/MapperByJson.java
@@ -1,9 +1,9 @@
-package ru.javarush.november.zaharin.thegame.services;
+package ru.javarush.november.zaharin.services;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import ru.javarush.november.zaharin.thegame.entity.Answer;
-import ru.javarush.november.zaharin.thegame.entity.Question;
+import ru.javarush.november.zaharin.entities.Answer;
+import ru.javarush.november.zaharin.entities.Question;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -18,7 +18,7 @@ public class MapperByJson {
                     new TypeReference<Map<Integer, Question>>() {
                     }));
         } catch (IOException | IllegalArgumentException e) {
-            throw new IllegalStateException("It doesn't work like that! Pay attention!");
+            throw new IllegalStateException("Error!");
         }
     }
 
@@ -28,7 +28,7 @@ public class MapperByJson {
                     new TypeReference<Map<Integer, Answer>>() {
                     }));
         } catch (IOException | IllegalArgumentException e) {
-            throw new IllegalStateException("It doesn't work like that! Pay attention!");
+            throw new IllegalStateException("Error!");
         }
     }
 }
